Move webgl utils out of ol/renderer/webgl/Layer module into their own module
